+++ R/esp-check-access.R
@@ -1,12 +1,12 @@
-#' Check access to SIANE data
+#' Check access to SIANE data resources
 #'
 #' @encoding UTF-8
 #' @keywords internal
 #' @description
-#' Check if **R** has access to resources at
+#' Check whether **R** has access to resources at
 #' <https://github.com/rOpenSpain/mapSpain/tree/sianedata>.
 #'
-#' @return A logical.
+#' @return Logical scalar, `TRUE` if accessible and `FALSE` otherwise.
 #'
 #' @seealso [giscoR::gisco_check_access()].
 #'
@@ -39,7 +39,7 @@ esp_check_access <- function() {
 
 #' Skip tests if SIANE data is not reachable
 #'
-#' @return invisible TRUE or skips the test
+#' @return Invisible. `TRUE` if offline logic passes or the test is skipped.
 #'
 #' @noRd
 skip_if_siane_offline <- function() {
@@ -62,7 +62,7 @@ skip_if_siane_offline <- function() {
 
 #' Skip tests if GISCO API is not reachable
 #'
-#' @return invisible TRUE or skips the test
+#' @return Invisible. `TRUE` if offline logic passes or the test is skipped.
 #'
 #' @noRd
 skip_if_gisco_offline <- function() {
@@ -84,7 +84,7 @@ skip_if_gisco_offline <- function() {
 }
 
 #' Internal function to check if we are on CRAN
-#' @return logical
+#' @return Logical scalar, `TRUE` if running on CRAN and `FALSE` otherwise.
 #' @noRd
 on_cran <- function() {
   env <- Sys.getenv("NOT_CRAN")