+++ R/esp-get-countries-siane.R
@@ -86,7 +86,7 @@ esp_get_countries_siane <- function(
 #' @param data_sf sf object
 #' @param country character vector of country codes or names
 #'
-#' @return sf object filtered
+#' @return An `sf` object with rows filtered by the provided country vector.
 #'
 #' @noRd
 filter_country <- function(data_sf, country = NULL) {