+++ NEWS.md
@@ -1,3 +1,5 @@
+# mapSpain (development version)
+
 # mapSpain 1.0.0
 
 This major release introduces a full overhaul of the codebase and test suite.