+++ NEWS.md
@@ -73,7 +73,7 @@ versions.
 
 # mapSpain 0.10.0
 
-- Fix a bug on `esp_get_prov_siane()` when selecting columns #115.
+- Fix a bug in `esp_get_prov_siane()` when selecting columns (#115).
 - New argument `type` in `esp_get_comarca()`. Now it is possible to extract
   different types of comarcas: Official (IGN), statistical (INE), agrarian or
   livestock comarcas (MAPA).